LayoutTests/fast/events/touch/touch-handler-count-expected.txt

This test checks that we correctly update the touch event handler count as event handlers are added and removed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test addEventListener/removeEventListener on the document.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 0
Test setting touch handlers on the document.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 4
PASS window.internals.touchEventHandlerCount(document) is 4
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 0
Test addEventListener/removeEventListener on the window.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 0
Test setting touch handlers on the window.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 4
PASS window.internals.touchEventHandlerCount(document) is 4
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 0
Test addEventListener/removeEventListener on a div.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 0
Test setting touch handlers on a div.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 4
PASS window.internals.touchEventHandlerCount(document) is 4
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 0
Test addEventListener/removeEventListener on a new div.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 0
Test setting touch handlers on a new div.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 3
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 2
PASS window.internals.touchEventHandlerCount(document) is 0
Test that nested Documents' touch handlers are properly removed from their parent Document.
PASS window.internals.touchEventHandlerCount(document) is 0
PASS window.internals.touchEventHandlerCount(document) is 1
PASS window.internals.touchEventHandlerCount(document) is 0
